requirements.txt
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.10.0
blis==1.3.0
catalogue==2.0.10
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
cloudpathlib==0.23.0
confection==0.1.5
cryptography==46.0.2
cymem==2.0.11
distro==1.9.0
Django==5.2.7
djangorestframework==3.16.1
groq==0.32.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
langcodes==3.5.0
language_data==1.3.0
marisa-trie==1.3.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
murmurhash==1.0.13
numpy==2.3.3
packaging==25.0
pdfminer.six==20250506
pdfplumber==0.11.7
pillow==11.3.0
preshed==3.0.10
psycopg==3.2.10
pycparser==2.23
pydantic==2.12.0
pydantic_core==2.41.1
Pygments==2.19.2
pypdfium2==4.30.0
python-decouple==3.8
requests==2.32.5
rich==14.1.0
setuptools==80.9.0
shellingham==1.5.4
smart_open==7.3.1
sniffio==1.3.1
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.5.3
srsly==2.5.1
thinc==8.3.6
tqdm==4.67.1
typer==0.19.2
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
wasabi==1.1.3
weasel==0.4.1
wrapt==1.17.3
